change position udf function in csv_manager file

+++ src/OC/zip_manager.py
@@ -14,7 +14,7 @@ class ZipIterator:
         self.bar = tqdm(total=len(self.all_files))
 
     def __iter__(self):
-        self.file = ""
+        self.all_files = ""
         return self
 
     def __next__(self):
@@ -30,6 +30,8 @@ class ZipIterator:
             print(f'taking 30 second of rest')
             time.sleep(30)
             self.all_files = self.open_zip()
+            if self.all_files is None:
+                raise StopIteration
             self.file = self.all_files.pop(0)
             self.bar = tqdm(total=len(self.all_files))
             self.bar.update(1)
@@ -40,12 +42,15 @@ class ZipIterator:
             raise StopIteration
 
     def open_zip(self):
-        dir_zip_name = self.list_zip.pop(0)
-        print(f'starting with another zip dir: {dir_zip_name}')
-        with zipfile.ZipFile(dir_zip_name, 'r') as zip_ref:
-            zip_ref.extractall(self.output_file)
-            zip_ref.close()
-        os.remove(dir_zip_name)
-        all_files = glob(os.path.join(self.output_file, '*.csv'))
+        if len(self.list_zip) > 0:
+            dir_zip_name = self.list_zip.pop(0)
+            print(f'starting with another zip dir: {dir_zip_name}')
+            with zipfile.ZipFile(dir_zip_name, 'r') as zip_ref:
+                zip_ref.extractall(self.output_file)
+                zip_ref.close()
+            os.remove(dir_zip_name)
+            all_files = glob(os.path.join(self.output_file, '*.csv'))
 
-        return all_files
+            return all_files
+        else:
+            return None
